requirements.txt
astunparse==1.6.3
brotlipy==0.7.0
certifi==2022.12.7
cloudpickle==2.2.1
contextlib2==21.6.0
docopt==0.6.2
et-xmlfile==1.1.0
flatbuffers==23.3.3
fonttools==4.25.0
gast==0.4.0
google-pasta==0.2.0
h5py==3.8.0
imageio==2.26.0
joblib==1.2.0
keras==2.11.0
libclang==15.0.6.1
llvmlite==0.40.1
mkl-fft==1.3.1
mkl-service==2.4.0
ml-collections==0.1.1
munkres==1.1.4
networkx==3.0
nibabel==5.0.1
numba==0.57.1
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
pandas==1.5.3
Pillow==9.3.0
ply==3.11
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydicom==2.3.1
pykwalify==1.8.0
PyQt5-sip==12.11.0
pyradiomics==3.0.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.1
shap==0.42.1
SimpleITK==2.2.1
slicer==0.0.7
tensorboard==2.11.2
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.27
torch==1.13.1+cu116
torchaudio==0.13.1
torchvision==0.14.1+cu116
wrapt==1.15.0
xgboost==1.7.4
